Just trying to add some kind of caching for the tickets
so that even if the bot restarts and theres still a ticket open, it doesnt lose the msgs sent in the channel.

What im trying to do right now is save the data into a file before the bot exits and then reading that file into the cache again to populate the var so the bot doesnt lose any msgs.